feat(web): add Mintlify to the logo marquee and split logo files
