+++ apps/web/src/components/landing/marquee-logos/databuddy-logo.tsx
@@ -1,28 +1,5 @@
 import type { SVGProps } from "react";
 
-export function InthLogo(props: SVGProps<SVGSVGElement>) {
-  return (
-    <svg
-      role="img"
-      viewBox="0 -2.556 120 41.6"
-      xmlns="http://www.w3.org/2000/svg"
-      {...props}
-    >
-      <title>Inth</title>
-      <path
-        clipRule="evenodd"
-        d="M119.966 32.178H64.195L57.126 38.999V-1.918h62.84zM61.52 2.694v5.904h9.302v19.036h5.901V8.598h9.369v-5.904z m29.21 0.034V27.634h5.903v-9.204h11.73v9.204h5.937V2.728h-5.937v9.833H96.633V2.728z"
-        fill="currentColor"
-        fillRule="evenodd"
-      />
-      <path
-        d="M24.82 2.63v5.869h-9.442v13.164h9.442v5.87H-0.001v-5.87h9.441V8.499H-0.001V2.63zM47.1 16.926V2.63h5.654v24.903h-4.111L34.863 13.156v14.377h-5.654V2.63h4.214z"
-        fill="currentColor"
-      />
-    </svg>
-  );
-}
-
 export function DatabuddyLogo(props: SVGProps<SVGSVGElement>) {
   return (
     <svg
@@ -90,64 +67,6 @@ export function DatabuddyLogo(props: SVGProps<SVGSVGElement>) {
       />
       <path
         d="M4611.76 752.35c-52.5-37.3-127.76 55.54-132.35 125.31-2.71 41.15 40.48 70.47 77.88 53.1 46.7-21.69 75.46-81.01 76.45-133.26.34-17.7-7.55-34.89-21.98-45.15Z"
-        fill="currentColor"
-      />
-    </svg>
-  );
-}
-
-export function EmdashLogo(props: SVGProps<SVGSVGElement>) {
-  return (
-    <svg
-      role="img"
-      viewBox="-0.453 -9.421 170 42.171"
-      xmlns="http://www.w3.org/2000/svg"
-      {...props}
-    >
-      <title>Emdash</title>
-      <path
-        d="M47.897 20.961V0.246H62.567V3.751H52.005V8.342H62.567V11.834H52.005V17.47H62.567V20.961H47.897Z"
-        fill="currentColor"
-      />
-      <path
-        d="M141.751 15.163C141.226 12.613 138.675 11.998 135.548 11.519 132.982 11.124 131.198 10.987 131.314 9.557 131.392 8.603 132.398 7.714 133.898 7.58 136.001 7.389 137.34 8.484 137.783 10.367L141.277 10.1C141.017 7.65 139.006 4.925 134.207 4.923 128.877 4.919 126.56 8.226 127.273 10.795 128.048 13.598 131.181 13.869 133.474 14.176 136.307 14.549 137.852 15.07 137.821 16.275 137.803 17.04 136.939 17.92 134.774 17.948 132.342 17.975 130.886 16.905 130.376 15.086L126.682 15.352C126.94 18.332 129.8 21.167 134.277 21.167 139.829 21.167 142.413 18.377 141.751 15.162V15.163Z"
-        fill="currentColor"
-      />
-      <path
-        d="M102.657 0.198V6.823C101.334 5.635 99.623 4.925 97.761 4.925 93.553 4.925 90.137 8.563 90.137 13.046 90.137 17.53 93.553 21.17 97.761 21.17 99.623 21.17 101.334 20.46 102.657 19.27L103.089 20.962H106.524V0.198H102.658 102.657ZM98.413 18.05C95.909 18.05 93.88 15.87 93.88 13.178 93.88 10.485 95.91 8.306 98.413 8.306 100.356 8.306 102.009 9.618 102.657 11.463 102.845 11.995 102.944 12.575 102.944 13.178 102.944 13.78 102.844 14.36 102.657 14.892 102.009 16.739 100.356 18.05 98.413 18.05Z"
-        fill="currentColor"
-      />
-      <path
-        d="M120.998 5.434V6.823C119.677 5.637 117.966 4.926 116.104 4.926 111.896 4.926 108.48 8.563 108.48 13.046 108.48 17.53 111.896 21.17 116.104 21.17 117.966 21.17 119.677 20.46 120.998 19.272L121.432 20.962H124.867V5.434H121.001 120.998ZM116.756 18.051C114.252 18.051 112.223 15.87 112.223 13.178 112.223 10.485 114.253 8.306 116.756 8.306 118.699 8.306 120.354 9.619 120.998 11.466 121.188 11.996 121.287 12.575 121.287 13.178 121.287 13.781 121.187 14.36 120.998 14.892 120.354 16.739 118.699 18.051 116.756 18.051Z"
-        fill="currentColor"
-      />
-      <path
-        d="M87.6 8.074C86.321 5.917 83.679 4.837 81.408 5.149 79.317 5.434 77.781 6.777 77.092 7.5 75.747 5.664 73.511 4.736 71.466 5.149 69.351 5.576 68.361 6.715 68.361 6.715L67.988 5.431H64.652V20.957H68.362V13.149C68.362 10.968 68.475 10.295 69.124 9.541 70.443 8.015 72.874 8.014 74.004 9.541 74.675 10.45 74.619 11.892 74.619 13.149V20.957C74.619 20.969 78.328 20.957 78.328 20.957V11.847C78.386 10.651 78.569 10.118 79.066 9.545 80.384 8.016 82.816 8.015 83.947 9.545 84.617 10.452 84.561 11.894 84.561 13.151V20.958C84.561 20.97 88.271 20.958 88.271 20.958V11.751C88.271 10.312 88.154 9.008 87.601 8.076L87.6 8.074Z"
-        fill="currentColor"
-      />
-      <path
-        d="M147.287 20.955H143.576V0.198H147.287V6.714C147.287 6.714 148.29 5.631 150.393 5.146 152.89 4.572 155.296 5.686 156.46 7.535 157.168 8.658 157.248 10.159 157.256 11.553 157.268 14.136 157.256 20.955 157.256 20.955 157.256 20.955 153.546 20.968 153.546 20.955V13.148C153.546 11.891 153.602 10.447 152.932 9.54 151.801 8.012 149.368 8.014 148.05 9.54 147.401 10.292 147.29 10.967 147.29 13.148V20.955H147.287Z"
-        fill="currentColor"
-      />
-      <path
-        d="M17.262 7.161H40.891L34.052 15.515H10.424L17.262 7.161Z"
-        fill="currentColor"
-      />
-    </svg>
-  );
-}
-
-export function StackAuthLogo(props: SVGProps<SVGSVGElement>) {
-  return (
-    <svg
-      role="img"
-      viewBox="-0.375 13.501 116 40.154"
-      xmlns="http://www.w3.org/2000/svg"
-      {...props}
-    >
-      <title>Stack Auth</title>
-      <path
-        d="M55.726 33.337q-2.03 0-3.665-0.683a6.96 6.69 0 0 1-2.678-1.975q-1.04-1.294-1.317-3.072l3.602-0.514q0.382 1.511 1.571 2.33 1.194 0.818 2.716 0.818 0.851 0 1.65-0.257a3.223 3.096 0 0 0 1.306-0.756q0.519-0.501 0.519-1.232 0-0.267-0.089-0.512a1.353 1.301 0 0 0-0.265-0.476 1.933 1.859 0 0 0-0.547-0.414 4.124 3.964 0 0 0-0.874-0.353l-4.744-1.341q-0.534-0.147-1.243-0.415a5.543 5.327 0 0 1-1.358-0.769 4.124 3.964 0 0 1-1.104-1.317q-0.431-0.828-0.431-2.085 0-1.768 0.925-2.938 0.928-1.17 2.475-1.743t3.425-0.561q1.89 0.025 3.374 0.622 1.483 0.597 2.487 1.742 1.001 1.134 1.445 2.781l-3.73 0.609a3.3 3.172 0 0 0-0.747-1.438 3.61 3.468 0 0 0-1.307-0.89 4.383 4.212 0 0 0-1.573-0.329 4.253 4.088 0 0 0-1.534 0.219q-0.711 0.232-1.168 0.683-0.445 0.452-0.444 1.073 0 0.574 0.368 0.94 0.367 0.353 0.926 0.572 0.558 0.218 1.142 0.366l3.17 0.83q0.711 0.181 1.574 0.487 0.864 0.292 1.648 0.828 0.798 0.524 1.308 1.39 0.519 0.866 0.519 2.194 0 1.415-0.62 2.476a5.413 5.204 0 0 1-1.651 1.744 7.733 7.434 0 0 1-2.358 1.025 10.31 9.911 0 0 1-2.703 0.341m17.325-0.366q-1.419 0.257-2.79 0.22-1.358-0.025-2.436-0.452a3.287 3.16 0 0 1-1.624-1.402 4.201 4.039 0 0 1-0.533-1.829 86.356 83.005 0 0 1-0.025-2.121V16.144h3.452v11.048q0 0.768 0.012 1.39 0.025 0.609 0.265 0.975 0.456 0.694 1.459 0.755a11.6 11.15 0 0 0 2.22-0.096zM63.311 22.362V19.803h9.742v2.559z m15.585 10.975q-1.469 0-2.498-0.536a3.996 3.841 0 0 1-1.546-1.451 4.292 4.126 0 0 1-0.523-2.013q0-0.914 0.293-1.669 0.29-0.756 0.938-1.329 0.659-0.586 1.765-0.976a12.89 12.389 0 0 1 1.813-0.475 41.244 39.644 0 0 1 2.383-0.391l2.931-0.427-1.242 0.659q0-1.17-0.584-1.719t-1.954-0.549q-0.76 0-1.585 0.353-0.826 0.353-1.154 1.256l-3.12-0.95a5.284 5.079 0 0 1 1.952-2.659q1.433-1.025 3.907-1.025 1.815 0 3.223 0.538a4.021 3.866 0 0 1 2.131 1.853q0.406 0.733 0.481 1.462 0.077 0.733 0.078 1.636v8.046h-3.021V30.263l0.432 0.561q-1.004 1.329-2.168 1.927-1.154 0.585-2.932 0.584m0.736-2.609q0.952 0 1.6-0.317 0.659-0.329 1.039-0.743 0.393-0.416 0.534-0.695 0.266-0.537 0.303-1.243 0.052-0.718 0.052-1.197l1.015 0.245q-1.533 0.244-2.489 0.414-0.95 0.159-1.533 0.292a9.023 8.672 0 0 0-1.027 0.294q-0.507 0.196-0.825 0.427a1.547 1.486 0 0 0-0.458 0.486 1.417 1.362 0 0 0-0.14 0.599q0 0.45 0.23 0.78 0.242 0.315 0.672 0.488 0.431 0.17 1.027 0.17m15.274 2.609q-2.131 0-3.653-0.913a6.278 6.034 0 0 1-2.335-2.489q-0.812-1.572-0.812-3.547 0-2 0.839-3.572a6.444 6.195 0 0 1 2.385-2.476q1.533-0.903 3.627-0.903 2.422 0 4.06 1.183 1.649 1.17 2.104 3.207l-3.45 0.878q-0.303-1.023-1.066-1.598a2.654 2.552 0 0 0-1.699-0.572q-1.09 0-1.789 0.513-0.696 0.501-1.028 1.377-0.329 0.866-0.33 1.963 0 1.719 0.786 2.793 0.798 1.06 2.361 1.06 1.168 0 1.776-0.513 0.608-0.513 0.914-1.463l3.527 0.707q-0.584 2.099-2.209 3.232-1.624 1.134-4.008 1.133m7.713-0.366l0.024-17.557h3.501v10.73l4.643-6.341h4.262L110.102 26.386l5.277 6.583h-4.49l-4.745-6.339v6.339zM32.489 21.709L18.205 14.911a2.629 2.527 0 0 0-2.331-0.003L-0.051 22.463v14.163l17.09 8.161 13.639-6.633v4.737l-13.639 6.581L-0.051 41.417v3.305c0 0.684 0.405 1.308 1.044 1.607l15.145 7.119a2.629 2.527 0 0 0 2.342-0.016l15.593-7.593v-11.217a1.754 1.685 0 0 0-2.514-1.519l-10.905 5.051v-4.477l13.419-6.528v-2.987a2.848 2.738 0 0 0-1.584-2.453M47.393 52.67l5.582-17.56h5.15l5.581 17.56H60.153l-4.972-15.487h0.659l-4.896 15.487z m3.348-3.659v-3.097h9.64v3.097zM70.065 53.06q-1.61 0-2.626-0.525-1.016-0.524-1.585-1.33a5.8 5.575 0 0 1-0.799-1.67 9.023 8.672 0 0 1-0.293-1.598 18.044 17.345 0 0 1-0.051-1.048v-7.389h3.501v6.219q0 0.45 0.052 1.159 0.052 0.694 0.316 1.401 0.266 0.709 0.864 1.182 0.61 0.476 1.712 0.476 0.445 0 0.951-0.133a2.384 2.292 0 0 0 0.952-0.513q0.443-0.391 0.721-1.133 0.293-0.756 0.293-1.975l1.978 0.901q0 1.56-0.658 2.926a5.413 5.204 0 0 1-1.992 2.208q-1.319 0.843-3.336 0.842m4.439-0.392v-4.365h-0.431v-8.803h3.476v13.17z m13.796 0q-1.42 0.257-2.792 0.22-1.358-0.025-2.436-0.452a3.287 3.16 0 0 1-1.622-1.401 4.201 4.039 0 0 1-0.533-1.829 85.067 81.766 0 0 1-0.025-2.122v-11.241h3.45v11.047q0 0.768 0.012 1.389 0.025 0.609 0.266 0.976 0.458 0.694 1.459 0.755a11.6 11.15 0 0 0 2.221-0.098z m-9.742-10.607v-2.561h9.742v2.561z m21.118 10.609v-6.22q0-0.45-0.053-1.146a5.413 5.204 0 0 0-0.317-1.415 2.77 2.664 0 0 0-0.876-1.182q-0.595-0.476-1.699-0.475-0.443 0-0.95 0.133a2.32 2.23 0 0 0-0.953 0.525q-0.445 0.379-0.734 1.122-0.279 0.743-0.279 1.976l-1.98-0.904q0-1.56 0.659-2.926a5.477 5.265 0 0 1 1.979-2.206q1.333-0.84 3.349-0.842 1.61 0 2.627 0.525 1.016 0.525 1.571 1.328 0.572 0.805 0.812 1.683 0.241 0.868 0.293 1.586 0.051 0.719 0.049 1.048v7.39z m-9.364 0V35.109h3.071v9.024h0.432v8.535z"
         fill="currentColor"
       />
     </svg>
