+++ apps/web/src/components/landing/marquee-logos/emdash-logo.tsx
@@ -1,0 +1,42 @@
+import type { SVGProps } from "react";
+
+export function EmdashLogo(props: SVGProps<SVGSVGElement>) {
+  return (
+    <svg
+      role="img"
+      viewBox="-0.453 -9.421 170 42.171"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <title>Emdash</title>
+      <path
+        d="M47.897 20.961V0.246H62.567V3.751H52.005V8.342H62.567V11.834H52.005V17.47H62.567V20.961H47.897Z"
+        fill="currentColor"
+      />
+      <path
+        d="M141.751 15.163C141.226 12.613 138.675 11.998 135.548 11.519 132.982 11.124 131.198 10.987 131.314 9.557 131.392 8.603 132.398 7.714 133.898 7.58 136.001 7.389 137.34 8.484 137.783 10.367L141.277 10.1C141.017 7.65 139.006 4.925 134.207 4.923 128.877 4.919 126.56 8.226 127.273 10.795 128.048 13.598 131.181 13.869 133.474 14.176 136.307 14.549 137.852 15.07 137.821 16.275 137.803 17.04 136.939 17.92 134.774 17.948 132.342 17.975 130.886 16.905 130.376 15.086L126.682 15.352C126.94 18.332 129.8 21.167 134.277 21.167 139.829 21.167 142.413 18.377 141.751 15.162V15.163Z"
+        fill="currentColor"
+      />
+      <path
+        d="M102.657 0.198V6.823C101.334 5.635 99.623 4.925 97.761 4.925 93.553 4.925 90.137 8.563 90.137 13.046 90.137 17.53 93.553 21.17 97.761 21.17 99.623 21.17 101.334 20.46 102.657 19.27L103.089 20.962H106.524V0.198H102.658 102.657ZM98.413 18.05C95.909 18.05 93.88 15.87 93.88 13.178 93.88 10.485 95.91 8.306 98.413 8.306 100.356 8.306 102.009 9.618 102.657 11.463 102.845 11.995 102.944 12.575 102.944 13.178 102.944 13.78 102.844 14.36 102.657 14.892 102.009 16.739 100.356 18.05 98.413 18.05Z"
+        fill="currentColor"
+      />
+      <path
+        d="M120.998 5.434V6.823C119.677 5.637 117.966 4.926 116.104 4.926 111.896 4.926 108.48 8.563 108.48 13.046 108.48 17.53 111.896 21.17 116.104 21.17 117.966 21.17 119.677 20.46 120.998 19.272L121.432 20.962H124.867V5.434H121.001 120.998ZM116.756 18.051C114.252 18.051 112.223 15.87 112.223 13.178 112.223 10.485 114.253 8.306 116.756 8.306 118.699 8.306 120.354 9.619 120.998 11.466 121.188 11.996 121.287 12.575 121.287 13.178 121.287 13.781 121.187 14.36 120.998 14.892 120.354 16.739 118.699 18.051 116.756 18.051Z"
+        fill="currentColor"
+      />
+      <path
+        d="M87.6 8.074C86.321 5.917 83.679 4.837 81.408 5.149 79.317 5.434 77.781 6.777 77.092 7.5 75.747 5.664 73.511 4.736 71.466 5.149 69.351 5.576 68.361 6.715 68.361 6.715L67.988 5.431H64.652V20.957H68.362V13.149C68.362 10.968 68.475 10.295 69.124 9.541 70.443 8.015 72.874 8.014 74.004 9.541 74.675 10.45 74.619 11.892 74.619 13.149V20.957C74.619 20.969 78.328 20.957 78.328 20.957V11.847C78.386 10.651 78.569 10.118 79.066 9.545 80.384 8.016 82.816 8.015 83.947 9.545 84.617 10.452 84.561 11.894 84.561 13.151V20.958C84.561 20.97 88.271 20.958 88.271 20.958V11.751C88.271 10.312 88.154 9.008 87.601 8.076L87.6 8.074Z"
+        fill="currentColor"
+      />
+      <path
+        d="M147.287 20.955H143.576V0.198H147.287V6.714C147.287 6.714 148.29 5.631 150.393 5.146 152.89 4.572 155.296 5.686 156.46 7.535 157.168 8.658 157.248 10.159 157.256 11.553 157.268 14.136 157.256 20.955 157.256 20.955 157.256 20.955 153.546 20.968 153.546 20.955V13.148C153.546 11.891 153.602 10.447 152.932 9.54 151.801 8.012 149.368 8.014 148.05 9.54 147.401 10.292 147.29 10.967 147.29 13.148V20.955H147.287Z"
+        fill="currentColor"
+      />
+      <path
+        d="M17.262 7.161H40.891L34.052 15.515H10.424L17.262 7.161Z"
+        fill="currentColor"
+      />
+    </svg>
+  );
+}
