+++ apps/web/src/components/landing/marquee-logos/inth-logo.tsx
@@ -1,0 +1,24 @@
+import type { SVGProps } from "react";
+
+export function InthLogo(props: SVGProps<SVGSVGElement>) {
+  return (
+    <svg
+      role="img"
+      viewBox="0 -2.556 120 41.6"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <title>Inth</title>
+      <path
+        clipRule="evenodd"
+        d="M119.966 32.178H64.195L57.126 38.999V-1.918h62.84zM61.52 2.694v5.904h9.302v19.036h5.901V8.598h9.369v-5.904z m29.21 0.034V27.634h5.903v-9.204h11.73v9.204h5.937V2.728h-5.937v9.833H96.633V2.728z"
+        fill="currentColor"
+        fillRule="evenodd"
+      />
+      <path
+        d="M24.82 2.63v5.869h-9.442v13.164h9.442v5.87H-0.001v-5.87h9.441V8.499H-0.001V2.63zM47.1 16.926V2.63h5.654v24.903h-4.111L34.863 13.156v14.377h-5.654V2.63h4.214z"
+        fill="currentColor"
+      />
+    </svg>
+  );
+}
