+++ apps/web/src/components/landing/marquee-logos/stack-auth-logo.tsx
@@ -1,0 +1,18 @@
+import type { SVGProps } from "react";
+
+export function StackAuthLogo(props: SVGProps<SVGSVGElement>) {
+  return (
+    <svg
+      role="img"
+      viewBox="-0.375 13.501 116 40.154"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <title>Stack Auth</title>
+      <path
+        d="M55.726 33.337q-2.03 0-3.665-0.683a6.96 6.69 0 0 1-2.678-1.975q-1.04-1.294-1.317-3.072l3.602-0.514q0.382 1.511 1.571 2.33 1.194 0.818 2.716 0.818 0.851 0 1.65-0.257a3.223 3.096 0 0 0 1.306-0.756q0.519-0.501 0.519-1.232 0-0.267-0.089-0.512a1.353 1.301 0 0 0-0.265-0.476 1.933 1.859 0 0 0-0.547-0.414 4.124 3.964 0 0 0-0.874-0.353l-4.744-1.341q-0.534-0.147-1.243-0.415a5.543 5.327 0 0 1-1.358-0.769 4.124 3.964 0 0 1-1.104-1.317q-0.431-0.828-0.431-2.085 0-1.768 0.925-2.938 0.928-1.17 2.475-1.743t3.425-0.561q1.89 0.025 3.374 0.622 1.483 0.597 2.487 1.742 1.001 1.134 1.445 2.781l-3.73 0.609a3.3 3.172 0 0 0-0.747-1.438 3.61 3.468 0 0 0-1.307-0.89 4.383 4.212 0 0 0-1.573-0.329 4.253 4.088 0 0 0-1.534 0.219q-0.711 0.232-1.168 0.683-0.445 0.452-0.444 1.073 0 0.574 0.368 0.94 0.367 0.353 0.926 0.572 0.558 0.218 1.142 0.366l3.17 0.83q0.711 0.181 1.574 0.487 0.864 0.292 1.648 0.828 0.798 0.524 1.308 1.39 0.519 0.866 0.519 2.194 0 1.415-0.62 2.476a5.413 5.204 0 0 1-1.651 1.744 7.733 7.434 0 0 1-2.358 1.025 10.31 9.911 0 0 1-2.703 0.341m17.325-0.366q-1.419 0.257-2.79 0.22-1.358-0.025-2.436-0.452a3.287 3.16 0 0 1-1.624-1.402 4.201 4.039 0 0 1-0.533-1.829 86.356 83.005 0 0 1-0.025-2.121V16.144h3.452v11.048q0 0.768 0.012 1.39 0.025 0.609 0.265 0.975 0.456 0.694 1.459 0.755a11.6 11.15 0 0 0 2.22-0.096zM63.311 22.362V19.803h9.742v2.559z m15.585 10.975q-1.469 0-2.498-0.536a3.996 3.841 0 0 1-1.546-1.451 4.292 4.126 0 0 1-0.523-2.013q0-0.914 0.293-1.669 0.29-0.756 0.938-1.329 0.659-0.586 1.765-0.976a12.89 12.389 0 0 1 1.813-0.475 41.244 39.644 0 0 1 2.383-0.391l2.931-0.427-1.242 0.659q0-1.17-0.584-1.719t-1.954-0.549q-0.76 0-1.585 0.353-0.826 0.353-1.154 1.256l-3.12-0.95a5.284 5.079 0 0 1 1.952-2.659q1.433-1.025 3.907-1.025 1.815 0 3.223 0.538a4.021 3.866 0 0 1 2.131 1.853q0.406 0.733 0.481 1.462 0.077 0.733 0.078 1.636v8.046h-3.021V30.263l0.432 0.561q-1.004 1.329-2.168 1.927-1.154 0.585-2.932 0.584m0.736-2.609q0.952 0 1.6-0.317 0.659-0.329 1.039-0.743 0.393-0.416 0.534-0.695 0.266-0.537 0.303-1.243 0.052-0.718 0.052-1.197l1.015 0.245q-1.533 0.244-2.489 0.414-0.95 0.159-1.533 0.292a9.023 8.672 0 0 0-1.027 0.294q-0.507 0.196-0.825 0.427a1.547 1.486 0 0 0-0.458 0.486 1.417 1.362 0 0 0-0.14 0.599q0 0.45 0.23 0.78 0.242 0.315 0.672 0.488 0.431 0.17 1.027 0.17m15.274 2.609q-2.131 0-3.653-0.913a6.278 6.034 0 0 1-2.335-2.489q-0.812-1.572-0.812-3.547 0-2 0.839-3.572a6.444 6.195 0 0 1 2.385-2.476q1.533-0.903 3.627-0.903 2.422 0 4.06 1.183 1.649 1.17 2.104 3.207l-3.45 0.878q-0.303-1.023-1.066-1.598a2.654 2.552 0 0 0-1.699-0.572q-1.09 0-1.789 0.513-0.696 0.501-1.028 1.377-0.329 0.866-0.33 1.963 0 1.719 0.786 2.793 0.798 1.06 2.361 1.06 1.168 0 1.776-0.513 0.608-0.513 0.914-1.463l3.527 0.707q-0.584 2.099-2.209 3.232-1.624 1.134-4.008 1.133m7.713-0.366l0.024-17.557h3.501v10.73l4.643-6.341h4.262L110.102 26.386l5.277 6.583h-4.49l-4.745-6.339v6.339zM32.489 21.709L18.205 14.911a2.629 2.527 0 0 0-2.331-0.003L-0.051 22.463v14.163l17.09 8.161 13.639-6.633v4.737l-13.639 6.581L-0.051 41.417v3.305c0 0.684 0.405 1.308 1.044 1.607l15.145 7.119a2.629 2.527 0 0 0 2.342-0.016l15.593-7.593v-11.217a1.754 1.685 0 0 0-2.514-1.519l-10.905 5.051v-4.477l13.419-6.528v-2.987a2.848 2.738 0 0 0-1.584-2.453M47.393 52.67l5.582-17.56h5.15l5.581 17.56H60.153l-4.972-15.487h0.659l-4.896 15.487z m3.348-3.659v-3.097h9.64v3.097zM70.065 53.06q-1.61 0-2.626-0.525-1.016-0.524-1.585-1.33a5.8 5.575 0 0 1-0.799-1.67 9.023 8.672 0 0 1-0.293-1.598 18.044 17.345 0 0 1-0.051-1.048v-7.389h3.501v6.219q0 0.45 0.052 1.159 0.052 0.694 0.316 1.401 0.266 0.709 0.864 1.182 0.61 0.476 1.712 0.476 0.445 0 0.951-0.133a2.384 2.292 0 0 0 0.952-0.513q0.443-0.391 0.721-1.133 0.293-0.756 0.293-1.975l1.978 0.901q0 1.56-0.658 2.926a5.413 5.204 0 0 1-1.992 2.208q-1.319 0.843-3.336 0.842m4.439-0.392v-4.365h-0.431v-8.803h3.476v13.17z m13.796 0q-1.42 0.257-2.792 0.22-1.358-0.025-2.436-0.452a3.287 3.16 0 0 1-1.622-1.401 4.201 4.039 0 0 1-0.533-1.829 85.067 81.766 0 0 1-0.025-2.122v-11.241h3.45v11.047q0 0.768 0.012 1.389 0.025 0.609 0.266 0.976 0.458 0.694 1.459 0.755a11.6 11.15 0 0 0 2.221-0.098z m-9.742-10.607v-2.561h9.742v2.561z m21.118 10.609v-6.22q0-0.45-0.053-1.146a5.413 5.204 0 0 0-0.317-1.415 2.77 2.664 0 0 0-0.876-1.182q-0.595-0.476-1.699-0.475-0.443 0-0.95 0.133a2.32 2.23 0 0 0-0.953 0.525q-0.445 0.379-0.734 1.122-0.279 0.743-0.279 1.976l-1.98-0.904q0-1.56 0.659-2.926a5.477 5.265 0 0 1 1.979-2.206q1.333-0.84 3.349-0.842 1.61 0 2.627 0.525 1.016 0.525 1.571 1.328 0.572 0.805 0.812 1.683 0.241 0.868 0.293 1.586 0.051 0.719 0.049 1.048v7.39z m-9.364 0V35.109h3.071v9.024h0.432v8.535z"
+        fill="currentColor"
+      />
+    </svg>
+  );
+}
